Adds play_buzzer for chime. Breaks up test_data tests
