+++ pyunifiprotect/data/__init__.py
@@ -11,6 +11,7 @@ from pyunifiprotect.data.devices import (
     Bridge,
     Camera,
     CameraChannel,
+    Chime,
     Doorlock,
     LCDMessage,
     Light,
@@ -33,6 +34,7 @@ from pyunifiprotect.data.types import (
     DEFAULT,
     DEFAULT_TYPE,
     ChimeDuration,
+    ChimeType,
     Color,
     CoordType,
     DoorbellMessageType,
@@ -42,6 +44,7 @@ from pyunifiprotect.data.types import (
     IRLEDMode,
     LightModeEnableType,
     LightModeType,
+    LockStatusType,
     ModelType,
     MountType,
     Percent,
@@ -70,7 +73,9 @@ __all__ = [
     "Bridge",
     "Camera",
     "CameraChannel",
+    "Chime",
     "ChimeDuration",
+    "ChimeType",
     "CloudAccount",
     "Color",
     "CoordType",
@@ -90,6 +95,7 @@ __all__ = [
     "LightModeEnableType",
     "LightModeType",
     "Liveview",
+    "LockStatusType",
     "ModelType",
     "MountType",
     "NVR",
